requirements.txt
streamlit>=1.32.0
openai>=1.12.0
anthropic>=0.18.1
mistralai==0.4.2
python-dotenv>=1.0.0
pandas>=2.0.0
pyyaml>=6.0.1
textstat>=0.7.3
pyspellchecker>=0.7.2 